Add static assert and create reset buffer pointer routine in event loop
[server] Add static assert against `struct cli_tcp_pkt`.

In order to ensure consistency across multiple machines, we have to make sure
that the size of `struct cli_tcp_pkt` is identical on any platform and
architecture. The reason behind this is because the struct will be sent over
network and read by other machine(s). If the struct is different in size or
alignment, it will cause a fatal problem for network I/O.

[server] Create reset buffer pointer routine in event loop.

Everytime the packet is received completely, we have to set the `recv_s` to zero
as we will overwrite the recv buffer. Fail to reset `recv_s` will lead to heap
overflow.

Acked-by: Louvian Lyndal <louvianlyndal@gmail.com>
Reviewed-by: Louvian Lyndal <louvianlyndal@gmail.com>
Signed-off-by: Ammar Faizi <ammarfaizi2@gmail.com>