Refactor tcp_io_uring.c (client and server)
I do not believe these are correct, so I want to delay the release.

I found something wrong with my logic about CQE handling. When I worked
with epoll, I can simply call `send()` and wait for it to return. But
in io_uring, it works differently. I have to handle the return value
later when the CQE has been generated. This improves the complexity
because I have to add more tracking to CQE for `send()` action as well
as `recv()` action.

There is more concern about this. SQE is limited, what should I do when
I run out of SQE? Abort the program? I am still considering it.

Signed-off-by: Ammar Faizi <ammarfaizi2@gmail.com>