main.c: use strcmp instead of strncmp for cmd comparison
Arthur Lapz <rlapz@gnuweeb.org> reported that he found a weird behavior
of teavpn2 command that it accepted a stupid input like "clientwkwkwkw"
as a valid command. This happened because the string comparison was
using `strncmp` which only compares only the first (at most) n bytes
of the input.

Fix for this is simply change the comparison function with `strcmp`.

This bug is trivial and harmless for users.

Link: https://t.me/GNUWeeb/458039
Reported-by: Arthur Lapz <rlapz@gnuweeb.org>
Signed-off-by: Ammar Faizi <ammarfaizi2@gnuweeb.org>