emerg: add src/emerg/arch/x64/emerg.o object to .PHONY list
The dependency check for this object is located in the sub Makefile
which is invoked in the sub shell. Thus the parent Makefile won't
notice the changes in its sub Makefile.

As such, add `$(BASE_DIR)/src/ext/hpc_emerg/src/emerg/arch/x64/emerg.o`
to the `.PHONY` list, let the sub Makefile handle the dependency check.

Signed-off-by: Ammar Faizi <ammarfaizi2@gmail.com>