server/linux/udp: perform read() syscall from tun_fd
When we open `/dev/net/tun` and apply `ioctl(TUNSETIFF)`, we can
add `IFF_MULTI_QUEUE` bit as flag which makes it possible to
open multiple file descriptors on the same virtual network interface.
All of these file descriptors can be used for parallel
`read()`/`write()`. So it's an ideal situation for multithreaded
implementation.

As we implement event loop for each thread, we need to identify
the file descriptors returned by `epoll_wait()`. In this case, we
only need to differentiate two types of fd:
  - TUN/TAP fd.
  - UDP socket fd.

Note: UDP socket only has single fd to `recv()` and `send()` from/to
everyone. So the following logic works very nice in hot path:

```
  if (fd == state->udp_fd) {
    // It must be UDP fd
    // Do recvfrom(fd), decapsulate the data, and write(tun_fd).
  } else {
    // It must be TUN fd
    // Do read(fd), encapsulate the data, and sendto() client(s).
  }
```

Signed-off-by: Ammar Faizi <ammarfaizi2@gmail.com>