udp_epoll.c: add reap_zombie_sessions
`reap_zombie_sessions` is a function to clean up zombie sessions.

This is mandatory for UDP socket as UDP is stateless, we don't
know whether the client has closed its connection or not. As such,
this commit implements timeout checking for every active session.

If a session is idle for more than `UDP_SESS_TIMEOUT` seconds, it
will get closed and the resource associated with it will get freed.

TODO:
The client should send close packet when it closes the connection.

Note:
Despite the client sends close packet, the server can't really rely
on it, because the packet may not arrive to the server, so this reap
function is still mandatory regardless that notification.

Signed-off-by: Ammar Faizi <ammarfaizi2@gmail.com>