server: close all clients before exit
Recall that UDP is stateless, if server doesn't notify the client
that it is going to exit, then the client will hang because the
client doesn't know whether the server is still online or not.

Fix for this is that, loop all over UDP sessions, if it is
authenticated, call `close_udp_session()`.

Signed-off-by: Ammar Faizi <ammarfaizi2@gmail.com>