server: don't treat repeated close packet as new connection
When the client sends `TCLI_PKT_CLOSE` more than once, the first
packet will remove the session entry from map.

However, the rest packet will initiate a new session again because the
server thinks that this is a new client (as it is not found in the map).

Fix for this is that, assume close packet as no-op if it is sent more
than once (only the first packet is processed).

Signed-off-by: Ammar Faizi <ammarfaizi2@gmail.com>