net/linux/iface.c: force delete stale route
If we have a stale route on the client app, delete it first.

In fact, we always perform the deletion on initialization (when
`override_default` is 1). But we don't care if there was no stale
route, because it does nothing if so (ignoring the error code of
`ip route delete` command is fully safe for this case).

Signed-off-by: Ammar Faizi <ammarfaizi2@gmail.com>