allocator: Implement page aligned allocator for hot buffer
Use `mmap()` directly for hot buffer allocations to make them be page
aligned.

The reason for page aligned buffers is to manage the cache-to-cache
transfers of these buffers, by hardware cache coherence mechanisms,
in SMP systems, so that their movement does not implicitly occur when
other variables are moved. This is for performance critical code,
where one expects contention in the access of variables by multiple
cpus (cores). The usual problem one tries to avoid, in this case, is
false sharing.

Signed-off-by: Ammar Faizi <ammarfaizi2@gmail.com>