build: Fix 32-bit build with `--gui` option
The library included for 32-bit build when it's invoked from 64-bit
environment was wrong. Fix this by conditionally select pkg-config
with 32-bit arch prefix if `--32-bit` option is specified.

This only adds support for x86 arch (`i686-linux-gnu-pkg-config`).
We can adjust this with different prefix if needed in the future.
If the target machine is the same with the host machine, it is not
a problem and should compile on any environment.

Signed-off-by: Alviro Iskandar Setiawan <alviro.iskandar@gmail.com>
Signed-off-by: Ammar Faizi <ammarfaizi2@gmail.com>