.github/workflows: Fix missing 32-bit dependencies
Build bot error again due to missing dependencies:
```
Some packages could not be installed. This may mean that you have
requested an impossible situation or if you are using the unstable
distribution that some required packages have not yet been created
or been moved out of Incoming.
The following information may help to resolve the situation:

The following packages have unmet dependencies:
Depends: gir1.2-gtk-3.0:i386 (= 3.24.20-0ubuntu1) but it is not going to be installed
Depends: libatk-bridge2.0-dev:i386 but it is not going to be installed
Depends: libatk1.0-dev:i386 (>= 2.15.1) but it is not going to be installed
Depends: libcairo2-dev:i386 (>= 1.14.0) but it is not going to be installed
Depends: libgdk-pixbuf2.0-dev:i386 (>= 2.40.0) but it is not going to be installed
Depends: libglib2.0-dev:i386 (>= 2.57.2) but it is not going to be installed
Depends: libpango1.0-dev:i386 (>= 1.41) but it is not going to be installed
Depends: dconf-gsettings-backend:i386 but it is not going to be installed or
gsettings-backend:i386
Depends: libgtk-3-0:i386 (= 3.24.20-0ubuntu1) but it is not going to be installed
E: Unable to correct problems, you have held broken packages.
Error: Process completed with exit code 100.
```

Fix this by installing all required dependencies.

For libgtk-3, they are:
```libgtk-3-common:i386 \
libgtk-3-dev:i386 \
gir1.2-gtk-3.0:i386 \
libatk-bridge2.0-dev:i386 \
libatk1.0-dev:i386 \
libcairo2-dev:i386 \
libgdk-pixbuf2.0-dev:i386 \
libglib2.0-dev:i386 \
libpango1.0-dev:i386 \
dconf-gsettings-backend:i386 \
libgtk-3-0:i386```

Please be green, this is tiring :(

Fixes: dc0f3eca60730795 (".github/workflows: Fix missing 32-bit build library")
Fixes: 30abce1d4fcba4c4 (".github/workflows: Add GUI build for GitHub build bot")
Signed-off-by: Alviro Iskandar Setiawan <alviro.iskandar@gmail.com>
Signed-off-by: Ammar Faizi <ammarfaizi2@gmail.com>