iface: Fix undefined behavior on `isspace()`
This patch addresses similar issue with the previous patch. But it is
caused by the wrong call handling to `isspace()`.

The argument of `isspace()` should be cast to `unsigned char` because
it will be undefined behavior if the value is not in range [0, 255].

How it be UB?
The `*p` is type of char, then it is cast to `unsigned`, the argument
prototype of `isspace()` is `int`, so it is still possible being a
negative value that results in UB.

Fix this by casting the argument to `unsigned char` to guarantee the
value is in [0, 255].

Cc: Khaerul Ilham <khaerulilham163@gmail.com>
Signed-off-by: Alviro Iskandar Setiawan <alviro.iskandar@gmail.com>
Link: https://t.me/TeaInside/47186
Signed-off-by: Ammar Faizi <ammarfaizi2@gmail.com>