common.h: Properly define cpu_relax() to avoid breaking the GUI code
A preparation patch to rework the build system. Previously, the
cpu_relax() macro comes from the HPC emergency library. But that library
is only compiled on the x86-64 architecture.

The GUI code inadvertently uses cpu_relax() without realizing the
breakage. Add a no-op macro to fix the build on archs other than x86-64
when buidling the GUI.

Signed-off-by: Ammar Faizi <ammarfaizi2@gnuweeb.org>