.github: Add multi architectures build to the GitHub build bot
Build multiple architectures with cross compilers to make sure we don't
break a specific architecture. The workflow script is lifted from
librung and modified to fit our needs.

Side note:
When cross-compiling on the GitHub workflow, the GTK+3 library is only
available for the x86_64 architecture. So, disable the GUI build for
other architectures.

Signed-off-by: Ammar Faizi <ammarfaizi2@gnuweeb.org>